requirements.txt
pillow==11.3.0
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
shiboken6==6.9.2
